fix(app): 미디어별 오류와 stage 재시도를 보존한다
